Adjust FB's CUnit tests for the new 64bit-compatible CUnit headers

+++ tests/fbcu/tests/fb-cunit-test.bas
@@ -1,14 +1,12 @@
 # include once "fbcu.bi"
 
-
-
 namespace fbcu.fbcu_test
 
-	function init cdecl () as integer
+	function init cdecl () as long
 		return 0
 	end function
 
-	function cleanup cdecl () as integer
+	function cleanup cdecl () as long
 		return 0
 	end function
